requirements.txt
numpy==1.26.0
typer==0.15.1
typing>=3.7.4.3
fastapi==0.115.6
streamlit
uvicorn==0.34.0
hydra-core==1.3.2
torch==2.5.1
torchsummary==1.5.1
torchvision>=0.20.1
invoke==2.2.0
pillow==11.0.0
scipy == 1.15.0
pytorch_lightning>=2.5.0.post0
wandb>=0.19.2
matplotlib>=3.4.3
requests
httpx
locust
pandas
onnx
onnxruntime
onnxscript
bentoml
google-cloud
google-cloud-run
google-cloud-storage
google-cloud-secret-manager
